Changed namespaces to org.kie.tools. Added a missing script for win32 support (powershell script).
